Improved mapping
IIUIC, this project contains a bunch of test-benches and also different
implementations of the same module for different technologies.  That
causes Sigasi to report errors about duplicate declarations of modules.
I've excluded a bunch of directories from the build.  Hopefully, that
makes the current state of the project consistent and removes the
duplicates.